refactor: step3 - 패키지 변경
