+++ src/main/java/domain/Person3.java
@@ -1,4 +1,4 @@
-package domain.step2;
+package domain;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
@@ -7,6 +7,8 @@ import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.Table;
 import jakarta.persistence.Transient;
+
+import java.util.Objects;
 
 @Table(name = "users")
 @Entity
@@ -27,4 +29,32 @@ public class Person3 {
 
     @Transient
     private Integer index;
+
+    public Person3() {
+    }
+
+    public Person3(Long id, String name, Integer age, String email) {
+        this.id = id;
+        this.name = name;
+        this.age = age;
+        this.email = email;
+    }
+
+    public Person3(String name, Integer age, String email) {
+        this.name = name;
+        this.age = age;
+        this.email = email;
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public boolean isIdValueNull() {
+        return Objects.isNull(this.id);
+    }
+
+    public void updateIdValue(long id) {
+        this.id = id;
+    }
 }
